feat: configure mac exclusive packages to only be installed when Darwin is detected

+++ skhd/install.sh
@@ -1,5 +1,8 @@
-mkdir -p "$HOME/.config/skhd"
-rm -rf $HOME/.config/skhd/*
-cp -rf $HOME/.dotfiles/skhd/* $HOME/.config/skhd
+if [ "$(uname -s)" == "Darwin" ]
+then
+	mkdir -p "$HOME/.config/skhd"
+	rm -rf $HOME/.config/skhd/*
+	cp -rf $HOME/.dotfiles/skhd/* $HOME/.config/skhd
 
-skhd --start-service
+	skhd --start-service
+fi
