+++ yabai/install.sh
@@ -1,5 +1,8 @@
+if [ "$(uname -s)" == "Darwin" ]
+then
 mkdir -p "$HOME/.config/yabai"
 rm -rf $HOME/.config/yabai/*
 cp -rf $HOME/.dotfiles/yabai/* $HOME/.config/yabai
 
 yabai --start-service
+fi
